requirements.txt
apache-airflow==2.7.3
pyspark==3.5.0
great-expectations==0.18.12
boto3==1.34.0
pandas==2.2.0
numpy==1.26.0
requests==2.31.0
sqlalchemy==2.0.23
ccxt==4.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.11
pytest==8.3.5